Expand ChannelsPresenter
