+++ packages/presentation/src/components/icons/Copy.svelte
