+++ packages/presentation/src/index.ts
@@ -22,3 +22,4 @@ export { default as Avatar } from './components/Avatar.svelte'
 export { default as MessageViewer } from './components/MessageViewer.svelte'
 export { default as AttributeEditor } from './components/AttributeEditor.svelte'
 export { default as Card } from './components/Card.svelte'
+export { default as Channels } from './components/Channels.svelte'
